requirements.txt
anthropic>=0.52.2
geopy>=2.4.1
google-generativeai>=0.8.5
matplotlib>=3.10.3
numpy>=2.2.6
opencv-python>=4.11.0.86
pandas>=2.3.0
pillow>=11.2.1
plotly>=6.1.2
reportlab>=4.4.1
requests>=2.32.3
scikit-learn>=1.7.0
scipy>=1.15.3
shapely>=2.1.1
streamlit>=1.45.1
